+++ navigation/lib-navigation/src/main/java/ru/surfstudio/android/navigation/command/fragment/RemoveUntil.kt
@@ -1,7 +1,7 @@
 package ru.surfstudio.android.navigation.command.fragment
 
 import ru.surfstudio.android.navigation.animation.Animations
-import ru.surfstudio.android.navigation.animation.DefaultAnimations
+import ru.surfstudio.android.navigation.animation.resource.NoResourceAnimations
 import ru.surfstudio.android.navigation.command.fragment.base.FragmentNavigationCommand
 import ru.surfstudio.android.navigation.route.fragment.FragmentRoute
 
@@ -14,7 +14,7 @@ import ru.surfstudio.android.navigation.route.fragment.FragmentRoute
  */
 data class RemoveUntil(
         override val route: FragmentRoute,
-        override val animations: Animations = DefaultAnimations.fragment,
+        override val animations: Animations = NoResourceAnimations,
         override var sourceTag: String = "",
         val isInclusive: Boolean
 ) : FragmentNavigationCommand