+++ navigation/RELEASE_NOTES.md
@@ -5,3 +5,7 @@
 * ANDDEP-312 Added navigation module
 * ANDDEP-1113 Fixed fragmentManager crash, on adding fragment when app is in background.
 * Fragment manager crash fix updated for Xiaomi on Android 9
+* ANDDEP-1110 Add removeAll and removeUntil fragment animations
+* Fix RemoveUntil command execution for single removal
+* Add methods to listen for changes of tab head in TabFragmentNavigator
+* Update removeUntil and removeAll animations: now they inherit animations from the first fragment of the removing bunch.